+++ app/src/main/java/com/nighthawkapps/wallet/android/NighthawkWalletApp.kt
@@ -1,12 +1,15 @@
 package com.nighthawkapps.wallet.android
 
 import android.app.Application
+import android.content.Context
 import androidx.camera.camera2.Camera2Config
 import androidx.camera.core.CameraXConfig
+import cash.z.ecc.android.sdk.ext.Twig
 import cash.z.ecc.android.sdk.ext.ZcashSdk
 import cash.z.ecc.android.sdk.type.ZcashNetwork
 import com.nighthawkapps.wallet.android.di.component.AppComponent
 import com.nighthawkapps.wallet.android.di.component.DaggerAppComponent
+import com.nighthawkapps.wallet.android.ext.StrictModeHelper
 
 class NighthawkWalletApp : Application(), CameraXConfig.Provider {
 
@@ -18,14 +21,30 @@ class NighthawkWalletApp : Application(), CameraXConfig.Provider {
     /** The amount of transparent funds that need to accumulate before autoshielding is triggered */
     val autoshieldThreshold: Long = ZcashSdk.ZATOSHI_PER_ZEC // 1 ZEC
 
-    override fun onCreate() {
+    override fun attachBaseContext(base: Context?) {
+        super.attachBaseContext(base)
+        // Setting a global reference to the application object is icky; we should try to refactor
+        // this away if possible.  Doing this in attachBaseContext instead of onCreate()
+        // to avoid any lifecycle issues, as certain components can run before Application.onCreate()
+        // (like ContentProvider initialization), but attachBaseContext will still run before that.
         instance = this
+    }
+
+    override fun onCreate() {
         super.onCreate()
+
+        // Register this before the uncaught exception handler, because we want to make sure the
+        // exception handler also doesn't do disk IO.  Since StrictMode only applies for debug builds,
+        // we'll also see the crashes during development right away and won't miss them if they aren't
+        // reported by the crash reporting.
+        if (BuildConfig.DEBUG) {
+            StrictModeHelper.enableStrictMode()
+            Twig.enabled(true)
+        }
+
         defaultNetwork = ZcashNetwork.from(resources.getInteger(R.integer.zcash_network_id))
         component = DaggerAppComponent.factory().create(this)
         component.inject(this)
-//        Set enabled to true to view SDK logs
-//        Twig.enabled(true)
     }
 
     override fun getCameraXConfig(): CameraXConfig {