+++ app/src/main/java/com/nighthawkapps/wallet/android/ui/history/HistoryFragment.kt
@@ -4,6 +4,7 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import androidx.lifecycle.lifecycleScope
+import androidx.paging.PagedList
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.nighthawkapps.wallet.android.R
@@ -85,8 +86,13 @@ class HistoryFragment : BaseFragment<FragmentHistoryBinding>() {
         transactions.size.let { newCount ->
             twig("got a new paged list of transactions of length $newCount")
             binding.groupEmptyViews.goneIf(newCount > 0)
-            val preSize = transactionAdapter.itemCount
-            transactionAdapter.submitList(transactions)
+            // tricky: we handle two types of lists, empty and PagedLists. It's not easy to construct an empty PagedList so the SDK currently returns an emptyList() but that will not cast to a PagedList
+            if (newCount == 0) {
+                transactionAdapter.submitList(null)
+            } else {
+                // tricky: for now, explicitly fail (cast exception) if the transactions are not in a PagedList. Otherwise, this would silently fail to show items and be hard to debug if we're ever passed a non-empty list that isn't an instance of PagedList. This awkwardness will go away when we switch to Paging3
+                transactionAdapter.submitList(transactions as PagedList<ConfirmedTransaction>)
+            }
         }
     }